Add Microsoft OAuth2 Functionality
Add Microsoft OAuth2 User Authentication.
This was pain in the ass. No documentation on how to use this or what scope is needed for accessing the sharepoint.
But its working like a charm. oauth is not yet a django app, the only file used is the API schema.
For reusability, someone gotta further work on this, thanks. It's probably me myself, idk.